fix a/ code review
