+++ project/package.json
@@ -55,7 +55,6 @@
     "@types/node": "^12.20.55",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
-    "@types/react-helmet": "^6.1.5",
     "@types/react-redux": "^7.1.24",
     "axios-mock-adapter": "^1.19.0",
     "eslint-config-htmlacademy": "^6.0.0",
