+++ project/src/components/browser-history.tsx
@@ -1,0 +1,5 @@
+import { createBrowserHistory } from 'history';
+
+const browserHistory = createBrowserHistory();
+
+export default browserHistory;
