--- project/src/components/header/header-user-no-auth.tsx
@@ -1,14 +1,0 @@
-import { AppRoute } from 'const/const';
-import { Link } from 'react-router-dom';
-
-export default function HeaderUserNoAuth() {
-  return (
-    <li className="header__nav-item user">
-      <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Login}>
-        <div className="header__avatar-wrapper user__avatar-wrapper">
-        </div>
-        <span className="header__login">Sign in</span>
-      </Link>
-    </li>
-  );
-}
