+++ project/src/components/header/user-no-auth.tsx
@@ -1,0 +1,18 @@
+import { AppRoute } from 'const/const';
+import { Link } from 'react-router-dom';
+
+export default function UserNoAuth() {
+  return (
+    <nav className="header__nav">
+      <ul className="header__nav-list">
+        <li className="header__nav-item user">
+          <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Login}>
+            <div className="header__avatar-wrapper user__avatar-wrapper">
+            </div>
+            <span className="header__login">Sign in</span>
+          </Link>
+        </li>
+      </ul>
+    </nav>
+  );
+}
