+++ project/src/components/history-route/history-route.tsx
@@ -1,0 +1,33 @@
+import {useState, useLayoutEffect} from 'react';
+import {Router} from 'react-router-dom';
+import type {BrowserHistory} from 'history';
+
+export type HistoryRouterProps = {
+  history: BrowserHistory
+  basename?: string
+  children?: React.ReactNode
+}
+
+export default function HistoryRouter({
+  basename,
+  children,
+  history,
+}: HistoryRouterProps) {
+  const [state, setState] = useState({
+    action: history.action,
+    location: history.location,
+  });
+
+  useLayoutEffect(() => history.listen(setState), [history]);
+
+  return (
+    <Router
+      basename={basename}
+      location={state.location}
+      navigationType={state.action}
+      navigator={history}
+    >
+      {children}
+    </Router>
+  );
+}
