+++ project/src/store/api-actions.ts
@@ -7,6 +7,7 @@ import { Offer } from 'types/offer';
 import { AppDispatch, State } from 'types/state';
 import { UserData } from 'types/user-data';
 import { loadOffers, redirectToRoute, requireAuthorization, setDataLoadedStatus } from './action';
+import { toast } from 'react-toastify';
 
 export const fetchOffersAction = createAsyncThunk<void, undefined, {
   dispatch: AppDispatch,
@@ -45,10 +46,16 @@ export const loginAction = createAsyncThunk<void, AuthData, {
 }>(
   'user/login',
   async ({email, password}, {dispatch, extra: api}) => {
-    const {data: {token}} = await api.post<UserData>(APIRoute.Login, {email, password});
-    saveToken(token);
-    dispatch(requireAuthorization(AuthorizationStatus.Auth));
-    dispatch(redirectToRoute(AppRoute.Main));
+    try {
+      const {data: {token}} = await api.post<UserData>(APIRoute.Login, {email, password});
+      saveToken(token);
+      dispatch(requireAuthorization(AuthorizationStatus.Auth));
+      dispatch(redirectToRoute(AppRoute.Main));
+    } catch {
+      toast.error('Something went wrong', {
+        position: toast.POSITION.TOP_CENTER
+      });
+    }
   },
 );
 
@@ -59,8 +66,14 @@ export const logoutAction = createAsyncThunk<void, undefined, {
 }>(
   'user/logout',
   async (_arg, {dispatch, extra: api}) => {
-    await api.delete(APIRoute.Logout);
-    dropToken();
-    dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
+    try {
+      await api.delete(APIRoute.Logout);
+      dropToken();
+      dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
+    } catch {
+      toast.error('Something went wrong', {
+        position: toast.POSITION.TOP_CENTER
+      });
+    }
   },
 );
