+++ project/src/store/index.ts
@@ -1,6 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { createApi } from 'services/api';
 import { reducer } from './reducer';
+import { redirect } from './middlewares/redirect';
 
 export const api = createApi();
 
@@ -11,5 +12,5 @@ export const store = configureStore({
       thunk: {
         extraArgument: api,
       },
-    }),
+    }).concat(redirect),
 });
