+++ project/src/utils/utils.ts
@@ -1,4 +1,4 @@
-import { Cities, MAX_RATING, SortingOptions } from 'const/const';
+import { MAX_RATING, SortingOptions } from 'const/const';
 import { Offer } from 'types/offer';
 
 export function convertRatingToPercent(rating: number): string {
@@ -29,6 +29,3 @@ export const sortOffers = (filteredOffers: Offer[], currentSortType: string) => 
       throw new Error(`${currentSortType} not exist`);
   }
 };
-
-export const cities = Object.keys(Cities);
-export const randomCity = cities[Math.floor(Math.random() * cities.length)];
